+++ db/migrate/006_add_index_to_checklist_items_assignee.rb
@@ -1,0 +1,5 @@
+class AddIndexToChecklistItemsAssignee < ActiveRecord::Migration[7.2]
+  def change
+    add_index :checklist_items, :assignee_id unless index_exists?(:checklist_items, :assignee_id)
+  end
+end